requirements.txt
paho-mqtt
numpy
prometheus_client
